+++ src/parser/map.c
@@ -1,0 +1,15 @@
+#include "parser.h"
+
+t_map	*get_map(char **file)
+{
+	t_map	*map;
+
+	map = malloc(sizeof(map));
+	if (set_lighting(file, map))
+		return (free(map), NULL);
+	if (set_camera(file, map))
+		return (free(map), NULL);
+	if (get_obj_arr(file, map))
+		return (free(map), NULL);
+	return (map);
+}
